+++ dynect/src/main/java/denominator/dynect/DynECTProvider.java
@@ -20,7 +20,7 @@ import dagger.Provides;
 import denominator.BasicProvider;
 import denominator.CheckConnection;
 import denominator.DNSApiManager;
-import denominator.QualifiedResourceRecordSetApi.Factory;
+import denominator.QualifiedResourceRecordSetApi;
 import denominator.ResourceRecordSetApi;
 import denominator.ZoneApi;
 import denominator.config.ConcatBasicAndQualifiedResourceRecordSets;
@@ -129,12 +129,9 @@ public class DynECTProvider extends BasicProvider {
             return new DynECTResourceRecordSetApi.Factory(api);
         }
 
-        @Provides
-        @Singleton
-        Map<Factory, Collection<String>> factoryToProfiles(GeoResourceRecordSetApi.Factory in) {
-            Map<Factory, Collection<String>> factories = new LinkedHashMap<Factory, Collection<String>>();
-            factories.put(in, Arrays.asList("geo"));
-            return factories;
+        @Provides(type = SET)
+        QualifiedResourceRecordSetApi.Factory factoryToProfiles(GeoResourceRecordSetApi.Factory in) {
+            return in;
         }
     }