+++ model/src/main/java/denominator/model/profile/Geo.java
@@ -2,7 +2,6 @@ package denominator.model.profile;
 
 import static denominator.common.Preconditions.checkArgument;
 import static denominator.common.Preconditions.checkNotNull;
-import static denominator.model.ResourceRecordSets.tryFindProfile;
 
 import java.beans.ConstructorProperties;
 import java.util.Collection;
@@ -29,7 +28,9 @@ public class Geo extends LinkedHashMap<String, Object> {
      * found.
      * 
      * @since 1.3.1
+     * @deprecated will be removed in version 4.0. use {@link #create(Map)}
      */
+    @Deprecated
     @SuppressWarnings("unchecked")
     public static Geo asGeo(Map<String, Object> profile) {
         if (profile == null)
@@ -55,9 +56,12 @@ public class Geo extends LinkedHashMap<String, Object> {
      * returns a Geo view of the {@code rrset} or null if no geo profile found.
      * 
      * @since 1.3.1
+     * @deprecated will be removed in version 4.0. use
+     *             {@link ResourceRecordSet#geo()}
      */
+    @Deprecated
     public static Geo asGeo(ResourceRecordSet<?> rrset) {
-        return asGeo(tryFindProfile(rrset, "geo"));
+        return rrset.geo();
     }
 
     /**